apps/api/pyproject.toml
[project]
name = "litestar-bb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "advanced-alchemy>=1.9.1",
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "argon2-cffi>=25.1.0",
    "email-validator>=2.3.0",
    "granian>=2.7.2",
    "litestar>=2.21.1",
    "psycopg[binary]>=3.3.3",
    "pydantic-settings>=2.13.1",
    "sqlalchemy>=2.0.48",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[project.scripts]
api = "app.cli:main"
api-dev = "app.cli:run-dev"

[tool.pyright]
include = ["main.py", "src", "tests"]
extraPaths = ["src"]
venvPath = "."
venv = ".venv"

[tool.ruff]
line-length = 100
target-version = "py313"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
